+++ CHANGELOG.md
@@ -47,19 +47,23 @@
 
 - Collidables work when used without ECS
 
-
-
 ## 27.10.2017
 
 ### Added
 
 - Support for static entities
+- Added SetID & GetID support for sprites
+- Added CreateIndicies for GraphicsDevice
+- Added Draw method for drawing array of sprites for GraphicsDevice
 
 ### Changed
 
 - Transform members are private
 - Transform stores status if it has been modified
-
+- Sprite now uses C/C++ Vertex array instead of std::array.
+- Vertex uv is now Vector3 instead of Vector2
+- Tilemap now draws all tiles using one draw call.
+- Tilemap now sets TileID for sprites
 
 
 ## 20.10.2017